// Backend/backend-api/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const AddressType = {
  DELIVERY: 'DELIVERY',
  INVOICE: 'INVOICE',
  STORE: 'STORE'
} as const

export type AddressType = (typeof AddressType)[keyof typeof AddressType]


export const DeliveryStatus = {
  PENDING: 'PENDING',
  IN_PROGRESS: 'IN_PROGRESS',
  COMPLETED: 'COMPLETED'
} as const

export type DeliveryStatus = (typeof DeliveryStatus)[keyof typeof DeliveryStatus]


export const ProductStatus = {
  ACTIVE: 'ACTIVE',
  INACTIVE: 'INACTIVE'
} as const

export type ProductStatus = (typeof ProductStatus)[keyof typeof ProductStatus]


export const SaleVariant = {
  UNIT: 'UNIT',
  BOX: 'BOX',
  PACK: 'PACK'
} as const

export type SaleVariant = (typeof SaleVariant)[keyof typeof SaleVariant]


export const UserRole = {
  WHOLESALER: 'WHOLESALER',
  RETAILER: 'RETAILER',
  SUPPORT: 'SUPPORT',
  DELIVERY: 'DELIVERY',
  WAREHOUSE: 'WAREHOUSE',
  ADMIN: 'ADMIN',
  SUPERADMIN: 'SUPERADMIN'
} as const

export type UserRole = (typeof UserRole)[keyof typeof UserRole]


export const UserStatus = {
  PENDING_VERIFICATION: 'PENDING_VERIFICATION',
  INACTIVE: 'INACTIVE',
  ACTIVE: 'ACTIVE',
  PENDING_REVIEW: 'PENDING_REVIEW',
  APPROVED: 'APPROVED',
  BANNED: 'BANNED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]